Revert "APIS-7086 (#850)"
This reverts commit d75b6339a528f56bb4b34041fa7d2d2e491ca5d6.
