+++ app/uk/gov/hmrc/thirdpartydeveloperfrontend/connectors/ThirdPartyDeveloperSessionConnector.scala
@@ -26,17 +26,13 @@ import uk.gov.hmrc.http.{HttpClient, SessionId => _, _}
 import uk.gov.hmrc.play.http.metrics.common.API
 
 import uk.gov.hmrc.apiplatform.modules.tpd.core.domain.models.SessionId
-import uk.gov.hmrc.apiplatform.modules.tpd.mfa.dto.AccessCodeAuthenticationRequest
 import uk.gov.hmrc.apiplatform.modules.tpd.session.domain.models._
-import uk.gov.hmrc.apiplatform.modules.tpd.session.dto.{SessionCreateWithDeviceRequest, UserAuthenticationResponse}
 import uk.gov.hmrc.thirdpartydeveloperfrontend.config.ApplicationConfig
-import uk.gov.hmrc.thirdpartydeveloperfrontend.domain._
 import uk.gov.hmrc.thirdpartydeveloperfrontend.domain.models.connectors._
 
 @Singleton
 class ThirdPartyDeveloperSessionConnector @Inject() (
     http: HttpClient,
-    encryptedJson: EncryptedJson,
     config: ApplicationConfig,
     metrics: ConnectorMetrics
   )(implicit val ec: ExecutionContext
@@ -44,38 +40,6 @@ class ThirdPartyDeveloperSessionConnector @Inject() (
 
   lazy val serviceBaseUrl: String = config.thirdPartyDeveloperSessionUrl
   val api: API                    = API("third-party-developer-session")
-
-  def authenticate(loginRequest: SessionCreateWithDeviceRequest)(implicit hc: HeaderCarrier): Future[UserAuthenticationResponse] = metrics.record(api) {
-    encryptedJson.secretRequest(
-      loginRequest,
-      http.POST[SecretRequest, ErrorOr[UserAuthenticationResponse]](s"$serviceBaseUrl/authenticate", _)
-    )
-      .map {
-        case Right(response)                                    => response
-        case Left(UpstreamErrorResponse(_, UNAUTHORIZED, _, _)) => throw new InvalidCredentials
-        case Left(UpstreamErrorResponse(_, FORBIDDEN, _, _))    => throw new UnverifiedAccount
-        case Left(UpstreamErrorResponse(_, LOCKED, _, _))       => throw new LockedAccount
-        case Left(UpstreamErrorResponse(_, NOT_FOUND, _, _))    => throw new InvalidEmail
-        case Left(err)                                          => throw err
-      }
-  }
-
-  def authenticateMfaAccessCode(
-      accessCodeAuthenticationRequest: AccessCodeAuthenticationRequest
-    )(implicit hc: HeaderCarrier
-    ): Future[UserSession] = metrics.record(api) {
-
-    encryptedJson.secretRequest(
-      accessCodeAuthenticationRequest,
-      http.POST[SecretRequest, ErrorOr[UserSession]](s"$serviceBaseUrl/authenticate-mfa", _)
-    )
-      .map {
-        case Right(response)                                   => response
-        case Left(UpstreamErrorResponse(_, BAD_REQUEST, _, _)) => throw new InvalidCredentials
-        case Left(UpstreamErrorResponse(_, NOT_FOUND, _, _))   => throw new InvalidEmail
-        case Left(err)                                         => throw err
-      }
-  }
 
   def updateSessionLoggedInState(sessionId: SessionId, request: UpdateLoggedInStateRequest)(implicit hc: HeaderCarrier): Future[UserSession] = metrics.record(api) {
     http.PUT[String, Option[UserSession]](s"$serviceBaseUrl/session/$sessionId/loggedInState/${request.loggedInState}", "")
