+++ it/test/uk/gov/hmrc/thirdpartydeveloperfrontend/connectors/ThirdPartyDeveloperSessionConnectorIntegrationSpec.scala
@@ -22,7 +22,6 @@ import org.scalatestplus.play.guice.GuiceOneAppPerSuite
 import play.api.http.Status._
 import play.api.inject.bind
 import play.api.inject.guice.GuiceApplicationBuilder
-import play.api.libs.json._
 import play.api.{Application, Configuration, Mode}
 import uk.gov.hmrc.http.{HeaderCarrier, UpstreamErrorResponse}
 
@@ -35,7 +34,6 @@ import uk.gov.hmrc.apiplatform.modules.tpd.session.domain.models.{LoggedInState,
 import uk.gov.hmrc.apiplatform.modules.tpd.session.dto._
 import uk.gov.hmrc.apiplatform.modules.tpd.test.builders.UserBuilder
 import uk.gov.hmrc.apiplatform.modules.tpd.test.utils.LocalUserIdTracker
-import uk.gov.hmrc.thirdpartydeveloperfrontend.domain._
 import uk.gov.hmrc.thirdpartydeveloperfrontend.domain.models.connectors._
 import uk.gov.hmrc.thirdpartydeveloperfrontend.utils.WireMockExtensions
 
@@ -68,202 +66,7 @@ class ThirdPartyDeveloperSessionConnectorIntegrationSpec extends BaseConnectorIn
     val mfaId: MfaId                                                     = MfaId.random
     val accessCodeAuthenticationRequest: AccessCodeAuthenticationRequest = AccessCodeAuthenticationRequest(userEmail, accessCode, nonce, mfaId)
 
-    val payloadEncryption: PayloadEncryption        = app.injector.instanceOf[PayloadEncryption]
-    val encryptedLoginRequest: JsValue              = Json.toJson(SecretRequest(payloadEncryption.encrypt(loginRequest).as[String]))
-    val encryptedTotpAuthenticationRequest: JsValue = Json.toJson(SecretRequest(payloadEncryption.encrypt(accessCodeAuthenticationRequest).as[String]))
-
     val underTest: ThirdPartyDeveloperSessionConnector = app.injector.instanceOf[ThirdPartyDeveloperSessionConnector]
-  }
-
-  "authenticate" should {
-    "return the session containing the user when the credentials are valid and MFA is disabled" in new Setup {
-
-      stubFor(
-        post(urlEqualTo("/authenticate"))
-          .withRequestBody(equalToJson(encryptedLoginRequest.toString))
-          .willReturn(
-            aResponse()
-              .withStatus(OK)
-              .withHeader("Content-Type", "application/json")
-              .withBody(s"""
-                           |{
-                           |  "accessCodeRequired": false,
-                           |  "mfaEnabled": false,
-                           |  "session": {
-                           |    "sessionId": "$sessionId",
-                           |    "loggedInState": "LOGGED_IN",
-                           |    "developer": {
-                           |      "userId":"$userId",
-                           |      "email":"${userEmail.text}",
-                           |      "firstName":"John",
-                           |      "lastName": "Doe",
-                           |      "registrationTime": "${nowAsText}",
-                           |      "lastModified": "${nowAsText}",
-                           |      "verified": true,
-                           |      "mfaDetails": [],
-                           |      "emailPreferences": { "interests" : [], "topics": [] }
-                           |    }
-                           |  }
-                           |}""".stripMargin)
-          )
-      )
-
-      val result: UserAuthenticationResponse = await(underTest.authenticate(loginRequest))
-
-      verify(1, postRequestedFor(urlMatching("/authenticate")).withRequestBody(equalToJson(encryptedLoginRequest.toString)))
-      result shouldBe UserAuthenticationResponse(
-        accessCodeRequired = false,
-        mfaEnabled = false,
-        session = Some(UserSession(sessionId, LoggedInState.LOGGED_IN, buildTrackedUser(userEmail)))
-      )
-    }
-
-    // "return the nonce when the credentials are valid and MFA is enabled" in new Setup {
-
-    //   stubFor(
-    //     post(urlEqualTo("/authenticate"))
-    //       .withRequestBody(equalToJson(encryptedLoginRequest.toString))
-    //       .willReturn(
-    //         aResponse()
-    //           .withStatus(OK)
-    //           .withHeader("Content-Type", "application/json")
-    //           .withBody(s"""
-    //                        |{
-    //                        |  "accessCodeRequired": true,
-    //                        |  "nonce": "$nonce"
-    //                        |}""".stripMargin)
-    //       )
-    //   )
-
-    //   val result: UserAuthenticationResponse = await(underTest.authenticate(loginRequest))
-
-    //   verify(1, postRequestedFor(urlEqualTo("/authenticate")).withRequestBody(equalToJson(encryptedLoginRequest.toString)))
-    //   result shouldBe UserAuthenticationResponse(accessCodeRequired = true, mfaEnabled = true, Some(nonce))
-    // }
-
-    "throw Invalid credentials when the credentials are invalid" in new Setup {
-      stubFor(
-        post(urlEqualTo("/authenticate"))
-          .withRequestBody(equalToJson(encryptedLoginRequest.toString))
-          .willReturn(
-            aResponse()
-              .withStatus(UNAUTHORIZED)
-              .withHeader("Content-Type", "application/json")
-          )
-      )
-
-      intercept[InvalidCredentials](await(underTest.authenticate(SessionCreateWithDeviceRequest(userEmail, userPassword, mfaMandatedForUser = Some(false), None))))
-    }
-
-    "throw LockedAccount exception when the account is locked" in new Setup {
-      stubFor(
-        post(urlEqualTo("/authenticate"))
-          .withRequestBody(equalToJson(encryptedLoginRequest.toString))
-          .willReturn(
-            aResponse()
-              .withStatus(LOCKED)
-              .withHeader("Content-Type", "application/json")
-          )
-      )
-
-      intercept[LockedAccount] {
-        await(underTest.authenticate(SessionCreateWithDeviceRequest(userEmail, userPassword, mfaMandatedForUser = Some(false), None)))
-      }
-    }
-
-    "throw UnverifiedAccount exception when the account is unverified" in new Setup {
-      stubFor(
-        post(urlEqualTo("/authenticate"))
-          .withRequestBody(equalToJson(encryptedLoginRequest.toString))
-          .willReturn(
-            aResponse()
-              .withStatus(FORBIDDEN)
-              .withHeader("Content-Type", "application/json")
-          )
-      )
-
-      intercept[UnverifiedAccount] {
-        await(underTest.authenticate(SessionCreateWithDeviceRequest(userEmail, userPassword, mfaMandatedForUser = Some(false), None)))
-      }
-    }
-
-    "fail on Upstream5xxResponse when the call return a 500" in new Setup {
-      stubFor(
-        post(urlEqualTo("/authenticate"))
-          .withRequestBody(equalToJson(encryptedLoginRequest.toString))
-          .willReturn(
-            aResponse()
-              .withStatus(INTERNAL_SERVER_ERROR)
-          )
-      )
-
-      intercept[UpstreamErrorResponse] {
-        await(underTest.authenticate(SessionCreateWithDeviceRequest(userEmail, userPassword, mfaMandatedForUser = Some(false), None)))
-      }.statusCode shouldBe INTERNAL_SERVER_ERROR
-    }
-  }
-
-  "authenticateMfaAccessCode" should {
-    "return the session containing the user when the accessCode and nonce are valid" in new Setup {
-      stubFor(
-        post(urlEqualTo("/authenticate-mfa"))
-          .withRequestBody(equalToJson(encryptedTotpAuthenticationRequest.toString))
-          .willReturn(
-            aResponse()
-              .withStatus(OK)
-              .withHeader("Content-Type", "application/json")
-              .withBody(s"""
-                           |{
-                           |  "sessionId": "$sessionId",
-                           |  "loggedInState": "LOGGED_IN",
-                           |    "developer": {
-                           |      "userId":"$userId",
-                           |      "email":"${userEmail.text}",
-                           |      "firstName":"John",
-                           |      "lastName": "Doe",
-                           |      "registrationTime": "${nowAsText}",
-                           |      "lastModified": "${nowAsText}",
-                           |      "verified": true,
-                           |      "mfaDetails": [],
-                           |      "emailPreferences": { "interests" : [], "topics": [] }
-                           |    }
-                           |}""".stripMargin)
-          )
-      )
-
-      val result: UserSession = await(underTest.authenticateMfaAccessCode(accessCodeAuthenticationRequest))
-
-      verify(1, postRequestedFor(urlMatching("/authenticate-mfa")).withRequestBody(equalToJson(encryptedTotpAuthenticationRequest.toString)))
-      result shouldBe UserSession(sessionId, LoggedInState.LOGGED_IN, buildTrackedUser(emailAddress = userEmail))
-    }
-
-    "throw Invalid credentials when the credentials are invalid" in new Setup {
-      stubFor(
-        post(urlEqualTo("/authenticate-mfa"))
-          .withRequestBody(equalToJson(encryptedTotpAuthenticationRequest.toString))
-          .willReturn(
-            aResponse()
-              .withStatus(BAD_REQUEST)
-              .withHeader("Content-Type", "application/json")
-          )
-      )
-
-      intercept[InvalidCredentials](await(underTest.authenticateMfaAccessCode(accessCodeAuthenticationRequest)))
-    }
-
-    "throw InvalidEmail when the email is not found" in new Setup {
-      stubFor(
-        post(urlEqualTo("/authenticate-mfa"))
-          .withRequestBody(equalToJson(encryptedTotpAuthenticationRequest.toString))
-          .willReturn(
-            aResponse()
-              .withStatus(NOT_FOUND)
-              .withHeader("Content-Type", "application/json")
-          )
-      )
-
-      intercept[InvalidEmail](await(underTest.authenticateMfaAccessCode(accessCodeAuthenticationRequest)))
-    }
   }
 
   "fetchSession" should {
