+++ test/uk/gov/hmrc/thirdpartydeveloperfrontend/controllers/endpointauth/EndpointScenarioSpec.scala
@@ -206,10 +206,10 @@ abstract class EndpointScenarioSpec extends AsyncHmrcSpec with GuiceOneAppPerSui
   when(tpdConnector.requestReset(*[LaxEmailAddress])(*)).thenReturn(Future.successful(OK))
   when(tpdConnector.getOrCreateUserId(*[LaxEmailAddress])(*)).thenReturn(Future.successful(UserId.random))
   when(tpdConnector.reset(*)(*)).thenReturn(Future.successful(OK))
-  when(tpdSessionConnector.authenticate(*)(*)).thenReturn(Future.successful(UserAuthenticationResponse(false, false, None, Some(session))))
+  when(tpdConnector.authenticate(*)(*)).thenReturn(Future.successful(UserAuthenticationResponse(false, false, None, Some(session))))
   when(tpdSessionConnector.fetchSession(eqTo(sessionId))(*)).thenReturn(Future.successful(session))
   when(tpdSessionConnector.deleteSession(eqTo(sessionId))(*)).thenReturn(Future.successful(OK))
-  when(tpdSessionConnector.authenticateMfaAccessCode(*)(*)).thenReturn(Future.successful(session))
+  when(tpdConnector.authenticateMfaAccessCode(*)(*)).thenReturn(Future.successful(session))
   when(tpdConnector.verify(*)(*)).thenReturn(Future.successful(OK))
   when(tpaProductionConnector.verify(*)(*)).thenReturn(Future.successful(ApplicationVerificationSuccessful))
   when(tpdConnector.resendVerificationEmail(*[LaxEmailAddress])(*)).thenReturn(Future.successful(OK))
